chore: update ns-3-alibabacloud submodule for binary rename

+++ ns-3-alibabacloud/simulation/scratch/decoupled_replay/fct_writer.h
@@ -44,11 +44,9 @@ inline void qp_finish(FILE *fout, Ptr<RdmaQueuePair> q) {
             (CustomHeader::GetStaticWholeHeaderSize() -
              IntHeader::GetStaticSize());
     uint64_t standalone_fct = base_rtt + total_bytes * 8000000000lu / b;
-    fprintf(fout, "%08x %08x %u %u %llu %lld %lld %llu\n", q->sip.Get(), q->dip.Get(),
-            q->sport, q->dport, (unsigned long long)q->m_size,
-            (long long)q->startTime.GetTimeStep(),
-            (long long)(Simulator::Now() - q->startTime).GetTimeStep(),
-            (unsigned long long)standalone_fct);
+    fprintf(fout, "%08x %08x %u %u %lu %lu %lu %lu\n", q->sip.Get(), q->dip.Get(),
+            q->sport, q->dport, q->m_size, q->startTime.GetTimeStep(),
+            (Simulator::Now() - q->startTime).GetTimeStep(), standalone_fct);
     fflush(fout);
 
     FlowTag flowTag;
